refactor(constants): split monolithic constants into dedicated modules
Break lib/constants.ts into focused modules under lib/constants/:
- site.constants.ts: SITE config, GITHUB_URL, SOCIALS
- projects.constants.ts: PROJECTS array
- resume.constants.ts: EDUCATION, EXPERIENCES, LANGUAGES, HOBBIES
- skills.constants.ts: SKILLS map
- terminal.constants.ts: CLOSE_MESSAGES
- guestbook.constants.ts: GUESTBOOK_CONFIG limits

Re-export everything from lib/constants/index.ts so existing
imports remain unchanged.
