+++ lib/constants/guestbook.constants.ts
@@ -1,0 +1,11 @@
+export const GUESTBOOK_CONFIG = {
+  DEFAULT_PAGE_SIZE: 25,
+  MAX_PAGE_SIZE: 100,
+  DEFAULT_RATE_LIMIT_MAX_PER_HOUR: 3,
+  MAX_NAME_LENGTH: 60,
+  MIN_NAME_LENGTH: 2,
+  MAX_MESSAGE_LENGTH: 300,
+  MIN_MESSAGE_LENGTH: 2,
+  MAX_WEBSITE_LENGTH: 240,
+  MAX_USER_AGENT_LENGTH: 255,
+} as const;
