+++ lib/constants/index.ts
@@ -1,0 +1,6 @@
+export { SITE, GITHUB_URL, SOCIALS } from "./site.constants";
+export { PROJECTS } from "./projects.constants";
+export { SKILLS } from "./skills.constants";
+export { EDUCATION, EXPERIENCES, LANGUAGES, HOBBIES } from "./resume.constants";
+export { CLOSE_MESSAGES } from "./terminal.constants";
+export { GUESTBOOK_CONFIG } from "./guestbook.constants";
