+++ lib/constants/projects.constants.ts
@@ -1,0 +1,80 @@
+export const PROJECTS = [
+  {
+    name: "Kabila App",
+    description:
+      "A Web3 social platform and NFT ecosystem built on Hedera, featuring an NFT marketplace, community plazas, launchpads and creator tools.",
+    url: "https://kabila.app",
+    imageUrl: "/projects/kabila-app.png",
+    tags: [
+      "Next.js",
+      "Tailwind",
+      "Typescript",
+      "Blockchain",
+      "Marketplace",
+      "Web3 Social",
+    ],
+  },
+  {
+    name: "Kabila Tools",
+    description:
+      "An all-in-one NFT toolkit for creators and communities on Hedera. A powerful and easy-to-use platform to create, launch, and manage your NFT collections while being in total control.",
+    url: "https://tools.kabila.app",
+    imageUrl: "/projects/kabila-tools.png",
+    tags: ["Next.js", "Tailwind", "Typescript", "Blockchain"],
+  },
+  {
+    name: "Kabila Wallet",
+    description:
+      "A non-custodial, simple, and secure wallet for the Hedera network. Manage your accounts, tokens, NFTs, and connect with DApps while maintaining full control of your assets.",
+    url: "https://kabila.app/docs/kabila-wallet",
+    imageUrl: "/projects/kabila-wallet.webp",
+    tags: ["Next.js", "Tailwind", "Javascript", "Blockchain"],
+  },
+  {
+    name: "Ferreira Borges Thomas",
+    description:
+      "A professional website for an automotive repair business, featuring services, transparent pricing and contact information.",
+    url: "https://thomas-ferreira.fr",
+    imageUrl: "/projects/fbt-auto-repair.png",
+    tags: ["Next.js", "Tailwind", "Typescript", "Freelance"],
+  },
+  {
+    name: "Brian Grav&Style",
+    description:
+      "A custom e-commerce platform for a professional engraving studio, featuring a product catalog, online ordering and an admin dashboard for order management.",
+    url: "https://brian-gravure.vercel.app",
+    imageUrl: "/projects/brian-gravure.png",
+    tags: ["Next.js", "Tailwind", "Typescript", "E-commerce"],
+  },
+  {
+    name: "Portfolio V1",
+    description:
+      "My previous portfolio website built with Next.js and Chakra UI, featuring a dark theme and smooth animations.",
+    url: "https://hakkaofdev-portfolio-v1.vercel.app",
+    imageUrl: "/projects/old-portfolio.png",
+    tags: ["Next.js", "Chakra UI", "Typescript", "Open-source"],
+  },
+  {
+    name: "ts-next-chakra-motion-kit",
+    description:
+      "Discover a starter kit which includes Next.js, Chakra-UI, Framer-Motion in Typescript. You have few components, Internationalization, SEO and more in this template.",
+    url: "https://ts-next-chakra-motion-kit.vercel.app",
+    imageUrl: "/projects/ts-next-chakra-motion-kit.png",
+    tags: ["Next.js", "Chakra UI", "Typescript", "Template", "Open-source"],
+  },
+  {
+    name: "AC-Vision",
+    description: "DASAN OLT/ONUs Supervision tool [Opensource].",
+    url: "https://github.com/hakkaofdev/ac-vision",
+    imageUrl: "/projects/ac-vision.png",
+    tags: ["Next.js", "Python", "Typescript", "Redis", "Docker", "Open-source"],
+  },
+  {
+    name: "RT'ransport",
+    description:
+      "A Percel Platform created for the project of my first year in DUT Networks & Telecoms (two-year university degree).",
+    url: "https://github.com/HakkaOfDev/RT-ransport",
+    imageUrl: "/projects/rt-ransport.png",
+    tags: ["Python", "Flask", "Tailwind", "Open-source"],
+  },
+];
