+++ lib/constants/resume.constants.ts
@@ -1,0 +1,94 @@
+export const EDUCATION = [
+  {
+    period: "2022-2023",
+    name: "System Administration and Security (vocational degree)",
+    location: "IUT Châlons (Châlons-en-Champagne, France)",
+  },
+  {
+    period: "2020-2022",
+    name: "DUT Networks and Telecommunications (technical degree)",
+    location: "IUT Châlons (Châlons-en-Champagne, France)",
+    descriptions: ["Salutatorian"],
+  },
+  {
+    period: "2019-2020",
+    name: "Scientific Baccalaureate (A level equivalent)",
+    location: "Lycée Jean Talon (Châlons-en-Champagne, France)",
+    descriptions: [
+      "Specialised in ISN (Computer and digital sicences)",
+      "With distinctions",
+    ],
+  },
+];
+
+export const EXPERIENCES = [
+  {
+    period: "Since July 2022",
+    name: "Software Engineer",
+    company: "kabila.app",
+    location: "Madrid, Spain",
+    descriptions: [
+      "Leading a team to deliver high-quality web applications",
+      "Implementing modern frontend architecture and best practices",
+      "Managing technical roadmap and sprint planning",
+      "Mentoring junior developers and conducting code reviews",
+      "Building reusable component libraries and design systems",
+      "Creating mobile apps with React Native and Expo",
+    ],
+  },
+  {
+    period: "Since July 2022",
+    name: "Freelance",
+    company: "Alexandre GOSSARD",
+    location: "Châlons-en-Champagne, France",
+    descriptions: [
+      "Developed full-stack web applications using React and Next.js",
+      "Built cross-platform mobile apps with React Native and Expo",
+      "Implemented CI/CD pipelines and automated deployment workflows",
+      "Provided technical consulting and architecture recommendations",
+      "Managed client relationships and project timelines independently",
+      "Delivering mobile apps for clients on the Google Play Store and Apple App Store",
+    ],
+  },
+  {
+    period: "September 2022 - August 2023",
+    name: "Apprenticeship",
+    company: "Arche MC2",
+    location: "Châlons-en-Champagne, France",
+    descriptions: [
+      "Implemented GitLab CI/CD pipelines for automated testing and deployment",
+      "Managed Docker containerization of applications and services",
+      "Automated system administration tasks using shell scripts",
+      "Monitored system health and implemented scalable solutions",
+    ],
+  },
+];
+
+export const LANGUAGES = [
+  {
+    lang: "🇫🇷 French",
+    level: "Native",
+  },
+  {
+    lang: "🇬🇧 English",
+    level: "C1",
+  },
+  {
+    lang: "🇷🇺 Russian",
+    level: "A2",
+  },
+  {
+    lang: "🇪🇸 Spanish",
+    level: "A2",
+  },
+];
+
+export const HOBBIES = [
+  "Motorcycles",
+  "Cars",
+  "Weightlifting",
+  "Calisthenics",
+  "Traveling",
+  "Manga",
+  "Gaming",
+];
