+++ lib/constants/terminal.constants.ts
@@ -1,0 +1,10 @@
+export const CLOSE_MESSAGES = [
+  "Nice try! You can't close me.",
+  "I'm not going anywhere. Try 'help' instead!",
+  "Nope. This terminal has trust issues with close buttons.",
+  "Error 418: I'm a teapot, not a closable window.",
+  "sudo rm -rf / ? Yeah... no.",
+  "You really thought that would work?",
+  "Close me? In this economy?",
+  "I'm built different. Try 'contact' instead.",
+];
